+++ releasenotes.md
@@ -109,6 +109,7 @@
 - Lambda check might run against a missing type definition if the function type alias was invalid.
 - Missing check when doing `$foo++` would crash the compiler if the variable wasn't initialized.
 - Incorrect handling of attribute operator symbols could crash the compiler instead of producing an error.
+- Crash instead of error when the first method parameter is a vaarg.
 
 ## 0.7.10 Change list